responding to PR comments
